backend/requirements.txt
psycopg2-binary
sqlalchemy
alembic
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
